+++ src/skagent/__init__.py
@@ -25,7 +25,7 @@ from .ann import Net, BlockPolicyNet
 from .simulation.monte_carlo import MonteCarloSimulator
 
 # Algorithm imports
-from .algos import best_response
+from .algos import tabular
 from .algos import vfi
 from .algos import maliar
 
@@ -50,7 +50,7 @@ __all__ = [
     # Simulation
     "MonteCarloSimulator",
     # Modules
-    "best_response",
+    "tabular",
     "vfi",
     "maliar",
     "utils",